Refactor component catalog into subcomponents
